+++ R/stan-params.R
@@ -34,6 +34,8 @@ fitted_pars_drm <- function(data_list) {
   }
   if (data_list$movement == 1)
     output <- c(output, "zeta")
+  if (data_list$rec_dd < 2)
+    output <- c(output, "beta")
   if (data_list$est_surv == 1)
     output <- c(output, "beta_s")
   if (data_list$ar_re > 0)
@@ -81,6 +83,8 @@ fitted_pars_lambda <- function(data_list) {
     output <- c(output, "beta_s")
   if (data_list$movement == 1)
     output <- c(output, "zeta")
+  if (data_list$rec_dd < 2)
+    output <- c(output, "beta")
   if (data_list$est_init == 1)
     output <- c(output, "log_init")
   return(output)