+++ NEWS.md
@@ -1,11 +1,10 @@
 # drmr 1.1.0
 
-- inclusion of density dependence!
+- added support for density dependence.
 
 - Minor bug in the `log_lik` function was corrected.
 
 - bug in the prediction function fixed.
-
 
 # drmr 1.0.1